requirements.txt
requests
json
load_dotenv
os